+++ .gitignore
@@ -7,10 +7,6 @@
 *.jar
 *.war
 *.ear
-*.iml
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
-
-#IntelliJ Files
-.idea/
