+++ README.md
@@ -2,25 +2,14 @@
 
 Shany Yeshanov and Chad Simmons
 
-Updated By: Brian Lim and Christopher Alsheikh
 
-
-## FlashCard Simulator to help study Flashcards. Includes:
-
-![Menu Screenshot](images/CreateDeck.png "Create a Deck Menu screenshot")
-
+FlashCard Simulator to help study Flashcards. Includes:
 - create cards
-- shuffle cards
+- shuffle cards 
 - flip cards to the back and front
 - view next card
-- save decks of cards
 
-## Modes:
-### Study mode: Review and flip cards to become familiar with the information on them.
-![Study Screenshot](images/Study.png "Quiz Menu")
-
-### Quiz mode: Review a deck of incorrect cards and keep a running number of how many were answered correct so far.
-![Quiz Screenshot](images/Quiz.png "Study Menu")
-
-
-
+AND ALSO:
+- save cards
+- quiz mode: review a deck of incorrect cards and keep a running number of how many were answered correct so far
+- the feedback from our second round of code reviews
