+++ build.xml
@@ -9,9 +9,9 @@
    <property name="webBaseUrl" value="http://www.cs.ucsb.edu/~${env.USER}/" />
 
   <property name="course" value="cs56" />
-  <property name="quarter" value="F17" />
+  <property name="quarter" value="W12" />
   <property name="issueNum" value="0000449" />
-  <property name="javaPkg" value="edu.ucsb.cs56.projects.utilities.flashcards" />
+  <property name="javaPkg" value="edu.ucsb.cs56.W12.syeshanov.flashcardsim" />
   <property name="mainClass" value="${javaPkg}.FlashCardApplication" />
   <property name="quizFrameClass" value="${javaPkg}.QuizFrame" />
   <property name="mainMenuFrameClass" value="${javaPkg}.MainMenuFrame" />
@@ -46,6 +46,24 @@
     </java>
   </target>
 
+  <target name="showquiz" depends="compile">
+    <java classname="${quizFrameClass}"  fork="true">
+       <classpath refid="project.class.path" />
+    </java>
+  </target>
+
+  <target name="showmain" depends="compile">
+    <java classname="${mainMenuFrameClass}"  fork="true">
+       <classpath refid="project.class.path" />
+    </java>
+  </target>
+
+  <target name="showcreate" depends="compile">
+    <java classname="${createDeckFrameClass}"  fork="true">
+       <classpath refid="project.class.path" />
+    </java>
+  </target>
+
  <target name="jws" depends="compile,jar">
 
    <delete dir="${webRoot}/${projectPath}/jws" />
@@ -61,6 +79,13 @@
    <echo>Java web start at ${jwsURL}</echo>
  </target>
 
+<target name="linenumbers" description="Creates a file containing all source files with line numbers">
+   <exec executable="bash">
+     <arg value="-c"/>
+     <arg value='tree -ifQ src | grep \.java\"$ | xargs pr -n > numberedsource.txt'/>
+   </exec>
+</target>
+
   <target name="clean" >
     <defaultexcludes remove="**/*~" />
     <delete>
@@ -70,6 +95,7 @@
     <delete dir="build" quiet="true" />
     <delete dir="javadoc" quiet="true" />
     <delete dir="dist" quiet="true" />
+    <delete dir="download" quiet="true" />
     <delete dir="temp" quiet="true" />
   </target>
 
@@ -128,7 +154,42 @@
    <signjar jar="dist/${projectName}.jar" alias="flashcardapp" keystore="keystore/flashcardapp.store" storepass="fl@sh!"/> 
  </target>
 
- <target name="publish" depends="jar,jws,javadoc">
+ <target name="download" >
+
+  <delete quiet="true">
+     <fileset dir="temp" includes="**/*" />
+     <fileset dir="download" includes="**/*" />
+  </delete>
+
+  <mkdir dir="download" />
+  <mkdir dir="temp" />
+  <mkdir dir="temp/${projectName}" />
+
+  <copy todir="temp/${projectName}">
+    <fileset dir="."
+     excludes="build/**, javadoc/**, **/*~, temp/**, dist/**, download/**"/>
+  </copy>
+
+  <tar destfile="temp/${projectName}.tar"
+       basedir="temp"
+       includes="${projectName}/**"
+   />
+
+  <gzip zipfile="download/${projectName}.tgz"
+        src="temp/${projectName}.tar" />
+
+  <zip destfile="download/${projectName}.zip"
+       basedir="temp"
+       includes="${projectName}/**"
+  />
+
+  <delete quiet="true">
+     <fileset dir="temp" includes="**/*"/>
+  </delete>
+
+</target>
+
+ <target name="publish" depends="jar,jws,javadoc,download,linenumbers">
    <mkdir dir="${webRoot}/${projectPath}" />
 
    <delete dir="${webRoot}/${projectPath}/browse" />
